# src/main/resources/static/openapi3/openapi3.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Demo User API",
    "description" : "사용자 관리 REST API - 테스트 코드로 자동 생성된 문서",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080"
  } ],
  "tags" : [ ],
  "paths" : {
    "/api/users" : {
      "get" : {
        "tags" : [ "User" ],
        "summary" : "모든 사용자 목록 조회",
        "description" : "모든 사용자 목록 조회",
        "operationId" : "users-get-all",
        "responses" : {
          "200" : {
            "description" : "200",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserList"
                },
                "examples" : {
                  "users-get-all" : {
                    "value" : "[ {\r\n  \"id\" : 1,\r\n  \"username\" : \"alice\",\r\n  \"email\" : \"alice@example.com\",\r\n  \"age\" : 25,\r\n  \"createdAt\" : null\r\n}, {\r\n  \"id\" : 2,\r\n  \"username\" : \"bob\",\r\n  \"email\" : \"bob@example.com\",\r\n  \"age\" : 30,\r\n  \"createdAt\" : null\r\n}, {\r\n  \"id\" : 3,\r\n  \"username\" : \"charlie\",\r\n  \"email\" : \"charlie@example.com\",\r\n  \"age\" : 35,\r\n  \"createdAt\" : null\r\n} ]"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "User" ],
        "summary" : "새 사용자 생성",
        "description" : "새 사용자 생성",
        "operationId" : "users-create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserCreateRequest"
              },
              "examples" : {
                "users-create" : {
                  "value" : "{\r\n  \"username\" : \"david\",\r\n  \"email\" : \"david@example.com\",\r\n  \"age\" : 28\r\n}"
                }
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "201",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                },
                "examples" : {
                  "users-create" : {
                    "value" : "{\r\n  \"id\" : 1,\r\n  \"username\" : \"david\",\r\n  \"email\" : \"david@example.com\",\r\n  \"age\" : 28,\r\n  \"createdAt\" : \"2025-10-23T17:52:47.1232819\"\r\n}"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users/filter" : {
      "get" : {
        "tags" : [ "User" ],
        "summary" : "최소 나이로 사용자 필터링",
        "description" : "최소 나이로 사용자 필터링",
        "operationId" : "users-filter",
        "parameters" : [ {
          "name" : "minAge",
          "in" : "query",
          "description" : "최소 나이",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "200",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserList"
                },
                "examples" : {
                  "users-filter" : {
                    "value" : "[ {\r\n  \"id\" : 3,\r\n  \"username\" : \"charlie\",\r\n  \"email\" : \"charlie@example.com\",\r\n  \"age\" : 35,\r\n  \"createdAt\" : null\r\n} ]"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users/search" : {
      "get" : {
        "tags" : [ "User" ],
        "summary" : "사용자명으로 검색 (부분 일치)",
        "description" : "사용자명으로 검색 (부분 일치)",
        "operationId" : "users-search",
        "parameters" : [ {
          "name" : "username",
          "in" : "query",
          "description" : "검색할 사용자명",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "200",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserList"
                },
                "examples" : {
                  "users-search" : {
                    "value" : "[ {\r\n  \"id\" : 1,\r\n  \"username\" : \"alice\",\r\n  \"email\" : \"alice@example.com\",\r\n  \"age\" : 25,\r\n  \"createdAt\" : null\r\n} ]"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}" : {
      "get" : {
        "tags" : [ "User" ],
        "summary" : "특정 ID의 사용자 조회",
        "description" : "특정 ID의 사용자 조회",
        "operationId" : "users-get-by-id",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "사용자 ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "200",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                },
                "examples" : {
                  "users-get-by-id" : {
                    "value" : "{\r\n  \"id\" : 1,\r\n  \"username\" : \"david\",\r\n  \"email\" : \"david@example.com\",\r\n  \"age\" : 28,\r\n  \"createdAt\" : \"2025-10-23T17:52:47.1232819\"\r\n}"
                  }
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "User" ],
        "summary" : "사용자 정보 전체 수정",
        "description" : "사용자 정보 전체 수정",
        "operationId" : "users-update",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "사용자 ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserUpdateRequest"
              },
              "examples" : {
                "users-update" : {
                  "value" : "{\r\n  \"email\" : \"alice.updated@example.com\",\r\n  \"age\" : 26\r\n}"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "200",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                },
                "examples" : {
                  "users-update" : {
                    "value" : "{\r\n  \"id\" : 1,\r\n  \"username\" : \"alice\",\r\n  \"email\" : \"alice.updated@example.com\",\r\n  \"age\" : 26,\r\n  \"createdAt\" : null\r\n}"
                  }
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "User" ],
        "summary" : "사용자 삭제",
        "description" : "사용자 삭제",
        "operationId" : "users-delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "삭제할 사용자 ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "204"
          }
        }
      },
      "patch" : {
        "tags" : [ "User" ],
        "summary" : "사용자 정보 부분 수정",
        "description" : "사용자 정보 부분 수정",
        "operationId" : "users-partial-update",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "사용자 ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserUpdateRequest"
              },
              "examples" : {
                "users-partial-update" : {
                  "value" : "{\r\n  \"email\" : \"bob.new@example.com\",\r\n  \"age\" : 31\r\n}"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "200",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                },
                "examples" : {
                  "users-partial-update" : {
                    "value" : "{\r\n  \"id\" : 2,\r\n  \"username\" : \"bob\",\r\n  \"email\" : \"bob.new@example.com\",\r\n  \"age\" : 31,\r\n  \"createdAt\" : null\r\n}"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "UserResponse" : {
        "title" : "UserResponse",
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "number",
            "description" : "사용자 ID"
          },
          "age" : {
            "type" : "number",
            "description" : "나이"
          },
          "email" : {
            "type" : "string",
            "description" : "이메일"
          },
          "username" : {
            "type" : "string",
            "description" : "사용자명"
          }
        }
      },
      "UserCreateRequest" : {
        "title" : "UserCreateRequest",
        "type" : "object",
        "properties" : {
          "age" : {
            "type" : "number",
            "description" : "나이"
          },
          "email" : {
            "type" : "string",
            "description" : "이메일"
          },
          "username" : {
            "type" : "string",
            "description" : "사용자명"
          }
        }
      },
      "UserUpdateRequest" : {
        "title" : "UserUpdateRequest",
        "type" : "object",
        "properties" : {
          "age" : {
            "type" : "number",
            "description" : "새 나이"
          },
          "email" : {
            "type" : "string",
            "description" : "새 이메일"
          }
        }
      },
      "UserList" : {
        "title" : "UserList",
        "type" : "array",
        "items" : {
          "type" : "object",
          "properties" : {
            "id" : {
              "type" : "number",
              "description" : "사용자 ID"
            },
            "age" : {
              "type" : "number",
              "description" : "나이"
            },
            "email" : {
              "type" : "string",
              "description" : "이메일"
            },
            "username" : {
              "type" : "string",
              "description" : "사용자명"
            }
          }
        }
      }
    }
  }
}